Add files via upload
Find the matroid intersection between matroids fromed by columns of S1 and S2 using Krogdahl Graph. Input formation: I=MatroidIntersection(E,S1,S2). E is the base set, option=[1,2] for the case S1 is a numerical matrix, and S2 is a structured matrix, which is the default option. For option=[1,1], S1 and S2 are both numerical matrices and you need to replace all the "sprank" by "rank". I is the intersection of the two matriods. The algorithm is due to a document of Dr. R. Chandrasekaran : http://www.utdallas.edu/~kam093020/papers/matroid-intersection.pdf.